Improve preset timer accessibility
