+++ TrackCount/Misc/TimeFormattable.swift
@@ -2,9 +2,27 @@
 //  TimeFormattable.swift
 //  TrackCount
 //
-//  Contains logic for formatting time integers to a human readable format.
+//  Contains logic for formatting time
 //
 
+import Foundation
+
+func accessibleTimeFormat(_ seconds: TimeInterval) -> String {
+    let total = Int(seconds)
+    let hours = total / 3600
+    let minutes = (total % 3600) / 60
+    let secs = total % 60
+    
+    let minuteString = "\(minutes) minute\(minutes == 1 ? "" : "s")"
+    let secondString = "\(secs) second\(secs == 1 ? "" : "s")"
+    
+    if hours > 0 {
+        let hourString = "\(hours) hour\(hours == 1 ? "" : "s")"
+        return "\(hourString), \(minuteString), and \(secondString)"
+    } else {
+        return "\(minuteString) and \(secondString)"
+    }
+}
 
 protocol TimeFormattable {
     func wholeSeconds() -> Int
